--- src/main/java/org/ject/support/domain/recruit/Recruit.java
@@ -1,23 +1,0 @@
-package org.ject.support.domain.recruit;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-import org.ject.support.domain.base.BaseTimeEntity;
-import org.ject.support.domain.member.JobFamily;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Recruit extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 20, nullable = false)
-    private String semester;
-
-    @Enumerated(EnumType.STRING)
-    @Column(columnDefinition = "varchar(45)", nullable = false)
-    private JobFamily jobFamily;
-}
